relocate sample event
