+++ src/inspect_ai/_eval/task/log.py
@@ -33,8 +33,8 @@ from inspect_ai.log._log import (
     eval_config_defaults,
 )
 from inspect_ai.log._recorders import Recorder
-from inspect_ai.log._recorders.buffer import SampleBufferDatabase, SampleEvent
-from inspect_ai.log._recorders.types import SampleSummary
+from inspect_ai.log._recorders.buffer import SampleBufferDatabase
+from inspect_ai.log._recorders.types import SampleEvent, SampleSummary
 from inspect_ai.log._transcript import Event
 from inspect_ai.model import (
     GenerateConfig,
