+++ src/inspect_ai/log/_recorders/buffer/__init__.py
@@ -1,11 +1,10 @@
 from .buffer import cleanup_sample_buffers, sample_buffer
-from .database import SampleBufferDatabase, SampleEvent
+from .database import SampleBufferDatabase
 from .types import AttachmentData, EventData, SampleBuffer, SampleData, Samples
 
 __all__ = [
     "AttachmentData",
     "EventData",
-    "SampleEvent",
     "SampleData",
     "Samples",
     "SampleBuffer",
