+++ src/inspect_ai/log/_recorders/types.py
@@ -1,7 +1,14 @@
 from pydantic import BaseModel, Field
 
+from inspect_ai.log._transcript import Event
 from inspect_ai.model._chat_message import ChatMessage
 from inspect_ai.scorer._metric import Score
+
+
+class SampleEvent(BaseModel):
+    id: str | int
+    epoch: int
+    event: Event
 
 
 class SampleSummary(BaseModel):
